requirements.txt
google-genai
crawl4ai
requests
pydantic
pydantic-settings
python-dotenv
playwright
